remove: 컨테이너 컴포넌트 중복으로 인해 삭제 후 import 변경
