+++ app/components/Board/Filter/FilterCategory/index.tsx
@@ -1,16 +1,21 @@
-"use client";
-import Filter from "@/components/Board/Filter";
-import Container from "@/components/Container";
-import { useState } from "react";
+'use client';
+import Filter from '@/components/Board/Filter';
+import { useState } from 'react';
 
 interface ComponentProps {
   filterItem: string[];
 }
 
 const FilterCategory = ({ filterItem }: ComponentProps) => {
-  const [selected, setSelected] = useState("전체");
+  const [selected, setSelected] = useState('전체');
 
-  return <Filter filterItem={filterItem} selected={selected} setSelected={setSelected} />;
+  return (
+    <Filter
+      filterItem={filterItem}
+      selected={selected}
+      setSelected={setSelected}
+    />
+  );
 };
 
 export default FilterCategory;
