+++ app/components/Container/index.tsx
@@ -1,12 +1,12 @@
-const Container = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <div
-      className="sm:mx-auto px-2 sm:px-6 lg:px-20
-    h-full w-full sm:w-[calc(100%_-_80px)] lg:w-[calc(100%_-_110px)]"
-    >
-      {children}
-    </div>
-  );
+import { FC, ReactNode } from 'react';
+
+interface ContainerProps {
+  children: ReactNode;
+  className?: string;
+}
+
+const Container: FC<ContainerProps> = ({ children, className }) => {
+  return <div className={`w-full md:max-w-3xl mx-auto px-5 ${className}`}>{children}</div>;
 };
 
 export default Container;
