+++ app/components/Heading/index.tsx
@@ -11,6 +11,7 @@ import {
   BackHeader,
   BackCloseHeader,
 } from './TypeHeader';
+import Container from '@/components/Container';
 
 // 헤더 유형별 컴포넌트 정의
 const HeaderElement = {
@@ -45,7 +46,11 @@ const HeadingComponent = () => {
 };
 
 const Heading = ({ children }: { children: ReactNode }) => {
-  return <header>{children}</header>;
+  return (
+    <header>
+      <Container>{children}</Container>
+    </header>
+  );
 };
 
 const Content = () => {
